libc: x86_64: be more specific about clobbers in memcpy

+++ libc/arch/x86_64/memset.c
@@ -2,7 +2,7 @@
 
 void * memset(void * dest, int c, size_t n) {
 	asm volatile("cld; rep stosb"
-	             : "=c"((int){0})
+	             : "=c"((unsigned long){0})
 	             : "rdi"(dest), "a"(c), "c"(n)
 	             : "flags", "memory", "rdi");
 	return dest;
